Remove customised header script
Not clear why we had a customised version of this file. An update to Blowfish
had changed the way that the mobile menu was controlled (additional JS),
which we were missing, hence the fact that the hamburger menu was inoperative.

Removing our specialisation and going back to the Blowfish default fixes this.